// src/app/models/response/profile-response.ts
export interface ProfileResponse {

  id: number;
  name: string;
  lastname: string;
  title: string;
  domicile: string;
  country: string;
  phone: string;
  province: string;
  city: string;
  whatsapp: string;
  specialty: string;
  mat_nac: string;
  mat_prov: string;
  presentation: string;

}